Fix submodule root not found
.git files in a submodule can be a file instead of of a directory.
The use of finddir means that within a submodule the .git of the
submodule will ignore resulting in tig being launch for the parent.
By trying to find a file fist we sort this problem.

It might be a problem in a case of nested submodules using files
and dirs.